wasm_gc: split from_mir.rs monolith into domain modules (#340)
The MIR body emitter had grown to 3139 lines — a second emitter monolith
alongside emit.rs. Split it into a `from_mir/` directory module by domain,
per peer review:

  mod.rs (796)          entry + the emit_mir_expr dispatcher + args helper
  pattern_match.rs (877) Match: primitive / Option / Result / List / variant
  builtins.rs (627)     Float/Int/Bool scalar + List + Vector + numeric BinOp
  strings.rs (205)      InterpolatedStr + String BinOp
  constructors.rs (172) Construct + Option/Result ctors
  coverage.rs (170)     --explain-mir-coverage predicate (diagnostic only)
  records.rs (139)      RecordCreate / RecordUpdate
  collections.rs (115)  List literals
  control.rs (100)      Try (`?`)

Pure code motion: functions moved verbatim, visibility bumped to
pub(crate), shared imports re-exported from mod.rs. No behavior change —
the dual byte-differential confirms it: single-file 365 fns and games
549 fns still render byte-identical (same counts, same bytes), 48-example
regression still validates, clippy -D warnings clean.
