+++ src/codegen/wasm_gc/body/from_mir/records.rs
@@ -1,0 +1,139 @@
+//! Record lowering: `MirExpr::RecordCreate` / `RecordUpdate`.
+//! Mirrors `emit_record_create` / `emit_record_update`.
+
+use super::*;
+
+/// Emit a record field / update value, mirroring `emit_record_create`'s
+/// per-field special-cases: an `Option.None` value emits through the
+/// constructor with the field's declared `T` (the bare-literal value's
+/// own `.ty()` may be a generic `Var`, so the field declaration is the
+/// authoritative shape), and an empty-list value emits `ref.null` of the
+/// field's declared `List<T>`. Everything else recurses via
+/// `emit_mir_expr`.
+pub(crate) fn emit_mir_record_field_value(
+    func: &mut Function,
+    value: &Spanned<MirExpr>,
+    decl_ty: &str,
+    slots: &SlotTable,
+    ctx: &EmitCtx<'_>,
+) -> Result<Option<()>, WasmGcError> {
+    if let MirExpr::Construct(c) = &value.node
+        && matches!(c.node.ctor, MirCtor::Builtin(BuiltinCtor::OptionNone))
+        && let Some(inner) = decl_ty
+            .trim()
+            .strip_prefix("Option<")
+            .and_then(|s| s.strip_suffix('>'))
+    {
+        return emit_mir_option_constructor(func, None, Some(inner.trim()), slots, ctx);
+    }
+    if let MirExpr::List(items) = &value.node
+        && items.is_empty()
+    {
+        let canonical: String = decl_ty.chars().filter(|c| !c.is_whitespace()).collect();
+        if let Some(list_idx) = ctx.registry.list_type_idx(&canonical) {
+            func.instruction(&Instruction::RefNull(wasm_encoder::HeapType::Concrete(
+                list_idx,
+            )));
+            return Ok(Some(()));
+        }
+    }
+    Ok(emit_mir_expr(func, value, slots, ctx)?.map(|_| ()))
+}
+
+/// Mirror of `emit_record_create` (emit.rs): a newtype record emits its
+/// single field's value directly; otherwise push every declared field
+/// (in declaration order) and `struct.new $type_idx`.
+pub(crate) fn emit_mir_record_create(
+    func: &mut Function,
+    type_name: &str,
+    fields: &[MirRecordField],
+    slots: &SlotTable,
+    ctx: &EmitCtx<'_>,
+) -> Result<Option<()>, WasmGcError> {
+    if ctx.registry.newtype_underlying(type_name).is_some() {
+        let field = fields.first().ok_or(WasmGcError::Validation(format!(
+            "newtype record `{type_name}` requires one field"
+        )))?;
+        return Ok(emit_mir_expr(func, &field.value, slots, ctx)?.map(|_| ()));
+    }
+    let type_idx = ctx
+        .registry
+        .record_type_idx(type_name)
+        .ok_or(WasmGcError::Validation(format!(
+            "unknown record type `{type_name}`"
+        )))?;
+    let decl_fields = ctx
+        .registry
+        .record_fields
+        .get(type_name)
+        .ok_or(WasmGcError::Validation(format!(
+            "record `{type_name}` missing field list"
+        )))?
+        .clone();
+    for (decl_name, decl_ty) in &decl_fields {
+        let provided =
+            fields
+                .iter()
+                .find(|f| &f.name == decl_name)
+                .ok_or(WasmGcError::Validation(format!(
+                    "record `{type_name}` missing field `{decl_name}`"
+                )))?;
+        if emit_mir_record_field_value(func, &provided.value, decl_ty, slots, ctx)?.is_none() {
+            return Ok(None);
+        }
+    }
+    func.instruction(&Instruction::StructNew(type_idx));
+    Ok(Some(()))
+}
+
+/// Mirror of `emit_record_update` (emit.rs): push each declared field in
+/// order — the override value when present, else `struct.get` it from
+/// the base — then `struct.new $type_idx`.
+pub(crate) fn emit_mir_record_update(
+    func: &mut Function,
+    type_name: &str,
+    base: &Spanned<MirExpr>,
+    updates: &[MirRecordField],
+    slots: &SlotTable,
+    ctx: &EmitCtx<'_>,
+) -> Result<Option<()>, WasmGcError> {
+    let type_idx = ctx
+        .registry
+        .record_type_idx(type_name)
+        .ok_or(WasmGcError::Validation(format!(
+            "unknown record type `{type_name}`"
+        )))?;
+    let decl_fields = ctx
+        .registry
+        .record_fields
+        .get(type_name)
+        .ok_or(WasmGcError::Validation(format!(
+            "record `{type_name}` missing field list"
+        )))?
+        .clone();
+    for (decl_name, decl_ty) in &decl_fields {
+        if let Some(override_field) = updates.iter().find(|f| &f.name == decl_name) {
+            if emit_mir_record_field_value(func, &override_field.value, decl_ty, slots, ctx)?
+                .is_none()
+            {
+                return Ok(None);
+            }
+        } else {
+            let field_idx = ctx
+                .registry
+                .record_field_index(type_name, decl_name)
+                .ok_or(WasmGcError::Validation(format!(
+                    "record `{type_name}` has no field `{decl_name}` to copy from base"
+                )))?;
+            if emit_mir_expr(func, base, slots, ctx)?.is_none() {
+                return Ok(None);
+            }
+            func.instruction(&Instruction::StructGet {
+                struct_type_index: type_idx,
+                field_index: field_idx,
+            });
+        }
+    }
+    func.instruction(&Instruction::StructNew(type_idx));
+    Ok(Some(()))
+}
